+++ configs/node.py
@@ -2,3 +2,6 @@ import os
 
 SKALE_VOLUME_PATH = '/skale_vol'
 NODE_DATA_PATH = '/skale_node_data'
+
+LOCAL_WALLET_FILENAME = 'local_wallet.json'
+LOCAL_WALLET_FILEPATH = os.path.join(NODE_DATA_PATH, LOCAL_WALLET_FILENAME)