+++ core/wallet.py
@@ -1,6 +1,20 @@
 import inspect
+
+from skale.utils.helper import private_key_to_address
+from web3 import Web3
+
 from core.config import URLS, LONG_LINE
 from core.helper import get_node_creds, construct_url, get_request
+from tools.helper import write_json
+from configs.node import LOCAL_WALLET_FILEPATH
+
+
+def set_wallet_by_pk(private_key):
+    address = private_key_to_address(private_key)
+    address_fx = Web3.toChecksumAddress(address)
+    local_wallet = {'address': address_fx, 'private_key': private_key}
+    write_json(LOCAL_WALLET_FILEPATH, local_wallet)
+    print(f'Local wallet updated: {local_wallet["address"]}')
 
 
 def get_wallet_info(config, format):